improved component structure
