+++ pages/_document.js
@@ -17,6 +17,12 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 
 class MyDocument extends Document {
+  // ray test touch <
+  // static async getInitialProps(ctx) {
+  //   const initialProps = await Document.getInitialProps(ctx);
+  //   return { ...initialProps };
+  // }
+  // ray test touch >
   render() {
     return (
       <Html>
